Event handler for removing treatment
Working on the event handler for removing specific treatments for an animal. I added a class to store properties for the analysis of that animal. This is to simplify the removal of data from the analysis.

Once the event handler is fully implemented for SecondList, removing things such as seizures, meals, etc will also be simple